refactor-47/비공개 데이터 저장 및 수정 로직 트랜ì 트랜잭션 적용 위해 계층 분리 close #47
