+++ triton_dashboard/build.gradle
@@ -49,9 +49,6 @@ dependencies {
 
 	// 응답 생성 렌더링을 위함
 	implementation 'org.springframework.boot:spring-boot-starter-webflux'
-
-	// 서버 껏다킬 필요 없이 .html 리컴파일 만으로도 변경사항 확인 가능
-	implementation 'org.springframework.boot:spring-boot-devtools'
 	
 	compileOnly 'org.projectlombok:lombok'
 	runtimeOnly 'com.h2database:h2'
