+++ triton_dashboard/src/main/java/com/triton/msa/triton_dashboard/private_data/dto/PrivateDataResponseDto.java
@@ -1,7 +1,5 @@
 package com.triton.msa.triton_dashboard.private_data.dto;
 
-import com.triton.msa.triton_dashboard.project.entity.Project;
-
 import java.time.Instant;
 
 public record PrivateDataResponseDto(
