+++ triton_dashboard/src/main/java/com/triton/msa/triton_dashboard/private_data/util/ZipExtractor.java
@@ -9,7 +9,6 @@ import org.apache.commons.io.FileUtils;
 import org.apache.tika.Tika;
 import org.apache.tika.exception.TikaException;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -27,7 +26,6 @@ public class ZipExtractor {
 
     private final Tika tika = new Tika();
 
-    @Transactional
     public List<ExtractedFile> extract(MultipartFile file, List<String> skipped) throws IOException {
         Path tempDir = Files.createTempDirectory("upload-zip");
